Extract IcebergNamespaceTestBase to prevent multiple test runs when TestIcebergNamespaceOperations is inherited by others.
