+++ iceberg/iceberg-rest-server/src/test/java/org/apache/gravitino/iceberg/service/rest/TestIcebergNamespaceOperations.java
@@ -18,17 +18,9 @@
  */
 package org.apache.gravitino.iceberg.service.rest;
 
-import com.google.common.collect.ImmutableMap;
 import java.util.Arrays;
-import java.util.List;
-import java.util.Map;
 import java.util.Optional;
-import java.util.stream.Collectors;
-import javax.ws.rs.client.Entity;
 import javax.ws.rs.core.Application;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
-import javax.ws.rs.core.Response.Status;
 import org.apache.gravitino.listener.api.event.Event;
 import org.apache.gravitino.listener.api.event.IcebergCreateNamespaceEvent;
 import org.apache.gravitino.listener.api.event.IcebergCreateNamespaceFailureEvent;
@@ -47,25 +39,13 @@ import org.apache.gravitino.listener.api.event.IcebergRegisterTablePreEvent;
 import org.apache.gravitino.listener.api.event.IcebergUpdateNamespaceEvent;
 import org.apache.gravitino.listener.api.event.IcebergUpdateNamespaceFailureEvent;
 import org.apache.gravitino.listener.api.event.IcebergUpdateNamespacePreEvent;
-import org.apache.iceberg.catalog.Namespace;
-import org.apache.iceberg.rest.requests.CreateNamespaceRequest;
-import org.apache.iceberg.rest.requests.ImmutableRegisterTableRequest;
-import org.apache.iceberg.rest.requests.RegisterTableRequest;
-import org.apache.iceberg.rest.requests.UpdateNamespacePropertiesRequest;
-import org.apache.iceberg.rest.responses.CreateNamespaceResponse;
-import org.apache.iceberg.rest.responses.GetNamespaceResponse;
-import org.apache.iceberg.rest.responses.ListNamespacesResponse;
-import org.apache.iceberg.rest.responses.UpdateNamespacePropertiesResponse;
 import org.glassfish.jersey.server.ResourceConfig;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-public class TestIcebergNamespaceOperations extends IcebergTestBase {
-
-  private final Map<String, String> properties = ImmutableMap.of("a", "b");
-  private final Map<String, String> updatedProperties = ImmutableMap.of("b", "c");
+public class TestIcebergNamespaceOperations extends IcebergNamespaceTestBase {
 
   private DummyEventListener dummyEventListener;
 
@@ -76,108 +56,6 @@ public class TestIcebergNamespaceOperations extends IcebergTestBase {
         IcebergRestTestUtil.getIcebergResourceConfig(
             MockIcebergNamespaceOperations.class, true, Arrays.asList(dummyEventListener));
     return resourceConfig;
-  }
-
-  private Response doCreateNamespace(String... name) {
-    CreateNamespaceRequest request =
-        CreateNamespaceRequest.builder()
-            .withNamespace(Namespace.of(name))
-            .setProperties(properties)
-            .build();
-    return getNamespaceClientBuilder()
-        .post(Entity.entity(request, MediaType.APPLICATION_JSON_TYPE));
-  }
-
-  private Response doRegisterTable(String tableName) {
-    RegisterTableRequest request =
-        ImmutableRegisterTableRequest.builder().name(tableName).metadataLocation("mock").build();
-    return getNamespaceClientBuilder(
-            Optional.of("register_ns"), Optional.of("register"), Optional.empty())
-        .post(Entity.entity(request, MediaType.APPLICATION_JSON_TYPE));
-  }
-
-  private Response doListNamespace(Optional<String> parent) {
-    Optional<Map<String, String>> queryParam =
-        parent.isPresent()
-            ? Optional.of(ImmutableMap.of("parent", parent.get()))
-            : Optional.empty();
-    return getNamespaceClientBuilder(Optional.empty(), Optional.empty(), queryParam).get();
-  }
-
-  private Response doUpdateNamespace(String name) {
-    UpdateNamespacePropertiesRequest request =
-        UpdateNamespacePropertiesRequest.builder()
-            .removeAll(Arrays.asList("a", "a1"))
-            .updateAll(updatedProperties)
-            .build();
-    return getUpdateNamespaceClientBuilder(name)
-        .post(Entity.entity(request, MediaType.APPLICATION_JSON_TYPE));
-  }
-
-  private Response doLoadNamespace(String name) {
-    return getNamespaceClientBuilder(Optional.of(name)).get();
-  }
-
-  private Response doNamespaceExists(String name) {
-    return getNamespaceClientBuilder(Optional.of(name)).head();
-  }
-
-  private Response doDropNamespace(String name) {
-    return getNamespaceClientBuilder(Optional.of(name)).delete();
-  }
-
-  private void verifyLoadNamespaceFail(int status, String name) {
-    Response response = doLoadNamespace(name);
-    Assertions.assertEquals(status, response.getStatus());
-  }
-
-  private void verifyLoadNamespaceSucc(String name) {
-    Response response = doLoadNamespace(name);
-    Assertions.assertEquals(Status.OK.getStatusCode(), response.getStatus());
-
-    GetNamespaceResponse r = response.readEntity(GetNamespaceResponse.class);
-    Assertions.assertEquals(name, r.namespace().toString());
-    Assertions.assertEquals(properties, r.properties());
-  }
-
-  protected void verifyDropNamespaceSucc(String name) {
-    Response response = doDropNamespace(name);
-    Assertions.assertEquals(Status.NO_CONTENT.getStatusCode(), response.getStatus());
-  }
-
-  private void verifyDropNamespaceFail(int status, String name) {
-    Response response = doDropNamespace(name);
-    Assertions.assertEquals(status, response.getStatus());
-  }
-
-  private void verifyNamespaceExistsStatusCode(int status, String name) {
-    Response response = doNamespaceExists(name);
-    Assertions.assertEquals(status, response.getStatus());
-  }
-
-  protected void verifyCreateNamespaceSucc(String... name) {
-    Response response = doCreateNamespace(name);
-    Assertions.assertEquals(Status.OK.getStatusCode(), response.getStatus());
-
-    CreateNamespaceResponse namespaceResponse = response.readEntity(CreateNamespaceResponse.class);
-    Assertions.assertTrue(namespaceResponse.namespace().equals(Namespace.of(name)));
-
-    Assertions.assertEquals(namespaceResponse.properties(), properties);
-  }
-
-  private void verifyRegisterTableSucc(String tableName) {
-    Response response = doRegisterTable(tableName);
-    Assertions.assertEquals(Status.OK.getStatusCode(), response.getStatus());
-  }
-
-  private void verifyRegisterTableFail(int statusCode, String tableName) {
-    Response response = doRegisterTable(tableName);
-    Assertions.assertEquals(statusCode, response.getStatus());
-  }
-
-  private void verifyCreateNamespaceFail(int statusCode, String... name) {
-    Response response = doCreateNamespace(name);
-    Assertions.assertEquals(statusCode, response.getStatus());
   }
 
   @Test
@@ -269,28 +147,6 @@ public class TestIcebergNamespaceOperations extends IcebergTestBase {
         dummyEventListener.popPostEvent() instanceof IcebergRegisterTableFailureEvent);
   }
 
-  private void dropAllExistingNamespace() {
-    Response response = doListNamespace(Optional.empty());
-    Assertions.assertEquals(Status.OK.getStatusCode(), response.getStatus());
-
-    ListNamespacesResponse r = response.readEntity(ListNamespacesResponse.class);
-    r.namespaces().forEach(n -> doDropNamespace(n.toString()));
-  }
-
-  private void verifyListNamespaceFail(Optional<String> parent, int status) {
-    Response response = doListNamespace(parent);
-    Assertions.assertEquals(status, response.getStatus());
-  }
-
-  private void verifyListNamespaceSucc(Optional<String> parent, List<String> schemas) {
-    Response response = doListNamespace(parent);
-    Assertions.assertEquals(Status.OK.getStatusCode(), response.getStatus());
-
-    ListNamespacesResponse r = response.readEntity(ListNamespacesResponse.class);
-    List<String> ns = r.namespaces().stream().map(n -> n.toString()).collect(Collectors.toList());
-    Assertions.assertEquals(schemas, ns);
-  }
-
   @ParameterizedTest
   @ValueSource(strings = {"", IcebergRestTestUtil.PREFIX})
   void testListNamespace(String prefix) {
@@ -307,22 +163,6 @@ public class TestIcebergNamespaceOperations extends IcebergTestBase {
     verifyListNamespaceSucc(Optional.of("list_foo3"), Arrays.asList("list_foo3.a", "list_foo3.b"));
 
     verifyListNamespaceFail(Optional.of("list_fooxx"), 404);
-  }
-
-  private void verifyUpdateNamespaceSucc(String name) {
-    Response response = doUpdateNamespace(name);
-    Assertions.assertEquals(Status.OK.getStatusCode(), response.getStatus());
-
-    UpdateNamespacePropertiesResponse r =
-        response.readEntity(UpdateNamespacePropertiesResponse.class);
-    Assertions.assertEquals(Arrays.asList("a"), r.removed());
-    Assertions.assertEquals(Arrays.asList("a1"), r.missing());
-    Assertions.assertEquals(Arrays.asList("b"), r.updated());
-  }
-
-  private void verifyUpdateNamespaceFail(int status, String name) {
-    Response response = doUpdateNamespace(name);
-    Assertions.assertEquals(status, response.getStatus());
   }
 
   @Test
